% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sensor_audit_summary.R
\name{print.sensor_audit_summary}
\alias{print.sensor_audit_summary}
\title{Print a sensor audit summary}
\usage{
\method{print}{sensor_audit_summary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{sensor_audit_summary}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the summary object.
}
\description{
Print a sensor audit summary
}
