% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sensor_audit.R
\name{summary.sensor_audit}
\alias{summary.sensor_audit}
\title{Summarize a sensor leakage audit}
\usage{
\method{summary}{sensor_audit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sensor_audit}.}

\item{...}{Additional arguments passed to the method.}
}
\value{
An object of class \code{sensor_audit_summary}.
}
\description{
Provides detailed summary information for a \code{sensor_audit}
object.
}
